requirements.txt
alembic==1.15.2
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.8
cryptography==44.0.3
dnspython==2.7.0
email_validator==2.2.0
Flask==2.2.5
Flask-Bcrypt==0.7.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pillow==11.2.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2025.2
qrcode==7.3.1
requests==2.26.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==1.26.20
Werkzeug==3.1.3
WTForms==3.2.1
